+++ Changelog.md
@@ -6,6 +6,7 @@
 ### Added
   - Generator HW:
     * Support generics in field length: field length in field is decided by max value, but MSBs above the parameter value are forced to 0 (suppose synthesis will properly simplify it)
+  - Generator RAL: new option "--ral_force_base" allows to force base class of derived register to still use uvm_reg
 
 ## [0.22.3] - 2026-05-08